feat: make sure we don't do things serially
