+++ features/step_definitions/echo_steps.rb
@@ -6,14 +6,14 @@ Given('we configure nonnative to manual') do
 
     config.definition do |d|
       d.process = 'features/support/bin/start 12_321'
-      d.timeout = 0.5
+      d.timeout = 5
       d.port = 12_321
       d.file = 'logs_12_321'
     end
 
     config.definition do |d|
       d.process = 'features/support/bin/start 12_322'
-      d.timeout = 0.5
+      d.timeout = 5
       d.port = 12_322
       d.file = 'logs_12_322'
     end
@@ -32,5 +32,6 @@ end
 
 Then('we should receive a {string} response') do |response|
   @responses.each { |r| expect(r).to eq(response) }
+ensure
   Nonnative.stop
 end
